requirements.txt
feedparser>=6.0
jinja2>=3.1
pyyaml>=6.0
markdown>=3.5
anthropic>=0.40
